Replace caching JAR FS provider functionality with Singleton FS cache
