+++ intellij-plugin-structure/tests/src/test/kotlin/com/jetbrains/plugin/structure/jar/SingletonCachingJarFileSystemProviderTest.kt
@@ -1,7 +1,6 @@
 package com.jetbrains.plugin.structure.jar
 
 import com.jetbrains.plugin.structure.base.utils.writeText
-import com.jetbrains.plugin.structure.xinclude.withSystemProperty
 import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
@@ -17,27 +16,29 @@ class SingletonCachingJarFileSystemProviderTest {
   val tempFolder = TemporaryFolder()
 
   @Test
-  fun `Jar Uris are normalized`() = withSystemProperty("com.jetbrains.plugin.structure.jar.SingletonCachingJarFileSystemProvider.retentionTime",0) {
+  fun `Jar Uris are normalized`() {
     val jarPath = tempFolder.root.toPath().resolve("test.jar")
     FileSystems.newFileSystem(URI.create("jar:${jarPath.toUri()}"), mapOf<String, Any>("create" to true)).use {
       it.getPath("hello.txt").writeText("Hello World")
     }
 
+    val fileSystemProvider = CachingJarFileSystemProvider(retentionTimeInSeconds = 0)
+
     //Getting two filesystems with different path that point to the same file.
-    val fs1 = SingletonCachingJarFileSystemProvider.getFileSystem(jarPath)
-    val fs2 = SingletonCachingJarFileSystemProvider.getFileSystem(
+    val fs1 = fileSystemProvider.getFileSystem(jarPath)
+    val fs2 = fileSystemProvider.getFileSystem(
       Path.of("${tempFolder.root.absolutePath}/../${tempFolder.root.name}/test.jar")
     )
 
     Assert.assertSame(fs1, fs2)
 
     //After jar is closed for the first time, filesystem remains open
-    SingletonCachingJarFileSystemProvider.close(jarPath)
+    fileSystemProvider.close(jarPath)
     Assert.assertTrue(fs1.isOpen)
     Assert.assertTrue(fs2.isOpen)
 
     //After closed for second time, the filesystem has no users and is closed
-    SingletonCachingJarFileSystemProvider.close(jarPath)
+    fileSystemProvider.close(jarPath)
     Assert.assertFalse(fs1.isOpen)
     Assert.assertFalse(fs2.isOpen)
   }
